02.12 - otthon fele magamtol fele Gorgi public repo, dankeschide <3

+++ 10 - LAMBDA and LINQ kifejezesek/LAMBDA and LINQ solton/Feladat - 04/Movie.cs
@@ -15,7 +15,7 @@ public class Movie
     /// USA bevétel
     /// </summary>
     [JsonPropertyName("US Gross")]
-    public int? USGross {get; set; }
+    public long? USGross {get; set; }
 
     /// <summary>
     /// Bevétel a világban
